requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Web scraping and crawling
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0

# AI
anthropic==0.18.1

# Billing
stripe==8.0.0

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0.1
click==8.1.7

# Task queue
celery==5.3.6
redis==5.0.1

# Reporting
weasyprint==60.2
jinja2==3.1.3
pandas==2.2.0

# Validation
validators==0.22.0
tldextract==5.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
